packet: Move message error tests into `message` module
